Pahihu's Memory manager and Object Logo fixes

+++ print.c
@@ -253,7 +253,7 @@ void real_print_node(FILE *strm, NODE *nd, int depth, int width) {
     } else if ((unsigned int)nd < 200) {    /* for debugging */
 	char num[] = "{small}    ";
 
-	sprintf(&num[7],"%d",nd);
+	sprintf(&num[7],"%d", (int)nd);
 	ndprintf(strm,num);
     } else if ((ndty = nodetype(nd)) & NT_PRIM) {
 	ndprintf(strm, "PRIM");
@@ -280,7 +280,7 @@ void real_print_node(FILE *strm, NODE *nd, int depth, int width) {
 	if (print_backslashes && (getarrorg(nd) != 1)) {
 	    char org[] = "@	";
 
-	    sprintf(&org[1],"%d",getarrorg(nd));
+	    sprintf(&org[1],"%ld",getarrorg(nd));
 	    ndprintf(strm,org);
 	}
     } else if (ndty == QUOTE) {
